voted event handler refactor
